refactor: pages/Test 폴더를 pages/Canvas로 이름 변경 및 /canvas 페이지 추가
Test라는 이름보다는 Canvas라는 이름이 더 적절하다고 판단해서 이름을
변경했습니다. 또한, 서버 연결 없이 화이트보드 테스트가 가능한
CanvasPage를 /canvas 경로에서 사용할 수 있도록 했습니다.
