+++ frontend/src/pages/Canvas/components/CanvasSection.tsx
