+++ frontend/src/pages/Canvas/components/ColorPanel.tsx
