+++ frontend/src/pages/Canvas/components/StickyNoteColorPanel.tsx
