+++ frontend/src/pages/Canvas/components/StickyNoteEditPanel.tsx
