+++ frontend/src/pages/Canvas/components/getStickyNoteInstance.ts
