+++ frontend/src/pages/Canvas/components/stateClickedQuestionContents.ts
