+++ frontend/src/pages/Canvas/components/stateIsQuestionListOpen.ts
