+++ frontend/src/pages/Canvas/components/stateStickyNoteEditPanelVisible.ts
