+++ frontend/src/pages/Canvas/components/stateStickyNoteInstance.ts
