+++ frontend/src/pages/Canvas/components/stateVideoRef.tsx
