+++ frontend/src/pages/Instructor/Instructor.tsx
@@ -1,5 +1,5 @@
 import Header from "@/components/Header/Header";
-import CanvasSection from "@/pages/Test/components/CanvasSection";
+import CanvasSection from "@/pages/Canvas/components/CanvasSection";
 
 const Instructor = () => {
   return (
